Invalid json causing errors while testing
The changes to this file made on April 19 2016 cause JS errors in UI testing. You are unable to use the dashboard UI due to bower errors parsing invalid json. The bracket at the beginning and the comma + brace at the end need to be removed so that the UI runs while using gulp serve to test changes.