Update setup.md
Changing the database setup from "use dashboarddb" to "use dashboard" and the output message from "switched to db dashboarddb" to "switched to dashboard". The db name is dashboard so this setup causes issues when setting up other properties files and calling the database dashboard and the dashboard user has no readWrite permissions to it.