Adding in support for multiple owners  (#2722)
* Allow remote API to take in multiple owners

Delegated null check to the API where it will
check to see if both the owner and owners field
is empty.

* Combine owner and owners request into one list

* Overwrite owners field with new entries

* Validate each owner being added

* Add owners to existing owner list in the dashboard

* Make sure that new owners list is unique

* Pass in validated new owners list instead of recomputing it

* Update dashboard constructors to take in List<Owner>

* Add testcase for multiple owners

* Added test for multiple owners and owner field

Takes into account repeats, within the owner and
owners field.  Updates TestUtil to load in multiple
users.

* Update DashboardService to call isUserValid

This allows the dashboard to save new users
if the current list has someone that has not
logged in before.

* Update owners check and owner creation

* Skip over invalid users during remote user validation

If the entire owners list is empty, the request is considered
invalid.